Structured code into module
