+++ src/myLocalStorage.js
@@ -1,0 +1,5 @@
+const getFromLocalSTorage = () =>
+  // eslint-disable-next-line implicit-arrow-linebreak
+  JSON.parse(localStorage.getItem('todos')) || [];
+
+export default getFromLocalSTorage;
